Dev2

+++ include/vq/representativeVector.h
@@ -1,9 +1,12 @@
+// SPDX-FileCopyrightText: 2024 Yuichiro Hatanaka
+// SPDX-License-Identifier: BSD-3-Clause
+
 #ifndef __REPRESENTATIVEVECTOR_H__
 #define __REPRESENTATIVEVECTOR_H__
 
 #include<stdio.h>
 #include<string.h>
 
-int representativeVector(FILE *block_fp, FILE *codebook_fp, FILE *vq_fp, unsigned char raw_vector_data[], int vector_element, unsigned short *representative_vector_number);
+int representativeVector(FILE *block_fp, FILE *codebook_fp, FILE *vq_fp, unsigned char raw_vector_data[], int vector_element, int *representative_vector_number);
 
 #endif
